Add an EstimateGeometricMean and implement it
This becomes the public API of our estimation methods, so we no longer export the direct functionality, and we update all tests to test the trait instead.  We also implement this for the exact method (technically it's just a _really_ good estimation), and add basic tests to confirm the two usages are identical.